requirements.txt
streamlit
pymongo
python-dotenv
pydantic
openai